:heavy_plus_sign: add: Shops -> Shop (NAILCASE-93)
