+++ src/main/java/com/nailcase/model/entity/Shop.java
@@ -1,5 +1,7 @@
 package com.nailcase.model.entity;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import com.nailcase.common.BaseEntity;
@@ -17,6 +19,7 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -53,16 +56,43 @@ public class Shop extends BaseEntity {
 	@Column(name = "address", length = 128)
 	private String address;
 
+	@Column(name = "available_seat")
+	private Integer availableSeats;
+
 	@OneToOne(mappedBy = "shop", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
 	private ShopInfo shopInfo;
 
 	@OneToMany(mappedBy = "shop", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
 	private Set<ShopHour> shopHours;
 
+	@Builder.Default
+	@OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY, mappedBy = "shop")
+	private List<NailArtist> nailArtistList = new ArrayList<>();
+
+	@Builder.Default
+	@OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY, mappedBy = "shop")
+	private List<Reservation> reservationList = new ArrayList<>();
+
+	@Builder.Default
+	@OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY, mappedBy = "shop")
+	private List<ReservationDetail> reservationDetailList = new ArrayList<>();
+
 	// shop service에 있는 메소드 그대로 사용
 	public void setOwnerId(Long userId) {
 		this.member = Member.builder()
 			.memberId(userId)
 			.build();
+	}
+
+	public void associateDown() {
+		this.nailArtistList.forEach(nailArtist -> nailArtist.associateDown(this));
+	}
+
+	public void minusAvailableSeats() {
+		this.availableSeats--;
+	}
+
+	public void plusAvailableSeats() {
+		this.availableSeats++;
 	}
 }
